scale nearest

+++ src/main.c
@@ -152,7 +152,7 @@ int main(int argc, char **argv) {
 
 else if ( strncmp( configuration.command, "scale_nearest", 11 ) == 0 ) {
     /* scale_nearest() function is defined in feature.h and implemented in feature.c */
-    float x = atoi(configuration.arguments[0]);
+    float x = atof(configuration.arguments[0]);
     scale_nearest(configuration.filenames[0], x);
   }
   return 0;
